pyproject.toml
[project]
name = "drf-clean-architecture"
version = "0.1.0"
description = "DRF Clean Architecture project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core
    "django>=5.1,<5.2",
    "djangorestframework>=3.15,<3.16",
    # Authentication (JWT)
    "djangorestframework-simplejwt>=5.3,<5.4",
    # API Documentation (Swagger / ReDoc)
    "drf-spectacular>=0.27,<0.28",
    # Filtering & Search
    "django-filter>=24.3,<25.0",
    # Database (PostgreSQL)
    "psycopg[binary]>=3.2,<4.0",
    # Environment Variables
    "python-decouple>=3.8,<4.0",
    # CORS Headers
    "django-cors-headers>=4.4,<5.0",
    # Redis Cache
    "django-redis>=5.4,<6.0",
    # Production Server
    "gunicorn>=22.0,<23.0",
    "whitenoise>=6.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.9",
    "factory-boy>=3.3",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
